Added health regeneration
Added health regeneration when player health percentage becomes lower than 5% and regenerates back to 20%. The health regenerates by 1% per 2 seconds. All values are customizable in the editor. Health regeneration percentage needs to be between 0 and 1 so I added a slider and tooltip to make it clear. Also added an error incase that percentage goes below 0 or above 1, just to make sure.